Implement StoredResourceController and enhance ResourceStorageService for filesystem storage support

+++ src/main/java/org/dataverse/marketplace/repository/StoredResourceRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoredResourceRepo extends JpaRepository<StoredResource, Integer> {
+public interface StoredResourceRepo extends JpaRepository<StoredResource, Long> {
 
 }
